Extract useTimeout hook
